Add no-bridge-import lint rule, move GlyphDraft to types/
New oxlint rule (shift/no-bridge-import) prevents app code from importing
@/bridge directly. Only Editor, Font, Glyph, testing, store, and commands
are allowed. GlyphDraft type moved from bridge/draft.ts to types/draft.ts
since it's part of the public API used by tool behaviors.
