+++ apps/desktop/src/renderer/src/lib/tools/select/behaviors/BendCurveBehaviour.ts
@@ -2,7 +2,7 @@ import { Vec2 } from "@shift/geo";
 import type { ToolContext } from "../../core/Behavior";
 import type { ToolEventOf } from "../../core/GestureDetector";
 import type { SelectHandlerBehavior, SelectState } from "../types";
-import type { GlyphDraft } from "@/bridge/draft";
+import type { GlyphDraft } from "@/types/draft";
 
 export class BendCurveBehaviour implements SelectHandlerBehavior {
   #draft: GlyphDraft | null = null;
