+++ scripts/oxlint/shift-plugin.mjs
@@ -552,5 +552,47 @@ export default {
         };
       },
     },
+
+    /**
+     * Ban direct imports from @/bridge in app code.
+     *
+     * The bridge is internal plumbing. App code should use Font and Glyph.
+     * Only Editor, Font class, testing, store, and commands may import bridge.
+     */
+    "no-bridge-import": {
+      meta: {
+        type: "suggestion",
+        messages: {
+          noBridge:
+            "Do not import from @/bridge directly. Use Font and Glyph instead. The bridge is internal.",
+        },
+        schema: [],
+      },
+      create(context) {
+        const filename = context.getFilename();
+
+        const BRIDGE_ALLOWED = [
+          "bridge/", // bridge internals can import each other
+          "Editor.ts", // orchestrator wires bridge to Font/Glyph
+          "lib/editor/Font.ts", // Font wraps bridge
+          "lib/model/glyph.ts", // Glyph wraps bridge
+          "testing/", // test helpers create bridge instances
+          "store/", // app store creates bridge
+          "commands/", // commands access bridge via context
+        ];
+
+        if (isAllowedFile(filename, BRIDGE_ALLOWED)) return {};
+        if (filename.includes(".test.")) return {};
+
+        return {
+          ImportDeclaration(node) {
+            const source = node.source && node.source.value;
+            if (typeof source === "string" && source.startsWith("@/bridge")) {
+              context.report({ node, messageId: "noBridge" });
+            }
+          },
+        };
+      },
+    },
   },
 };
